refactor: move decodeData to its own file
